IR-505: Refactoring: Moved graphviz conversion
Move the conversion logic (to build the DOT format) into the
'conversion' module where it belongs.
