+++ server/data/incidentTypeConfiguration/conversion.ts
@@ -12,7 +12,6 @@ import {
 /**
  * Converts a NOMIS incident type configuration into the new DPS one
  */
-// eslint-disable-next-line import/prefer-default-export
 export function fromNomis(nomisConfig: NomisIncidentTypeConfiguration): DpsIncidentTypeConfiguration {
   const sortByQuestionSequence = (q1: NomisQuestionConfiguration, q2: NomisQuestionConfiguration) =>
     q1.questionSeq - q2.questionSeq
@@ -49,6 +48,38 @@ export function fromNomis(nomisConfig: NomisIncidentTypeConfiguration): DpsIncid
       return qs
     }, {}),
   }
+}
+
+/**
+ * Converts a DPS incident type into the graphviz/DOT format
+ */
+export function toGraphviz(config: DpsIncidentTypeConfiguration): string {
+  const questions = Object.values(config.questions)
+
+  let result = `digraph ${config.incidentType} {`
+  // Left to right
+  result += '  rankdir=LR;\n'
+  // Nodes are cicles
+  result += '  node [shape = circle];\n'
+  // Start arrow to first question
+  result += '  START_NODE [label="", shape=none];\n'
+  result += `  START_NODE -> ${config.startingQuestionId} [label = "start"];\n`
+  // End node
+  result += '  END_NODE [label="END", shape="doublecircle"];\n'
+
+  // Adds the questions as nodes, answers as edges
+  for (const question of questions) {
+    result += `  ${question.id} [label = "${question.label}"];\n`
+    for (const answer of question.answers) {
+      const nextNode = answer.nextQuestionId ?? 'END_NODE'
+
+      result += `  ${question.id} -> ${nextNode} [label = "${answer.label}"];\n`
+    }
+  }
+
+  result += '}'
+
+  return result
 }
 
 function typeFromNomisCode(nomisCode: NomisType): Type {
